move serial and dma serial to own file
