+++ src/io/mod.rs
@@ -23,7 +23,10 @@
 //!
 //! This module contains implementations of helper macros for print and println.
 
-pub use self::imp::*;
+mod serial;
+
+pub use io::serial::imp::*;
+pub use self::dma::*;
 
 #[cfg(not(feature="serial"))]
 mod imp {
@@ -35,228 +38,3 @@ mod imp {
     }
 }
 
-#[cfg(feature="serial")]
-mod imp {
-    use altos_core::volatile::Volatile;
-    use altos_core::syscall::sleep;
-    use altos_core::sync::{Mutex, CriticalSection};
-    use altos_core::collections::RingBuffer;
-    use core::fmt::{self, Write, Arguments};
-    use peripheral::usart::{UsartX, Usart, USART2_TX_CHAN, USART2_RX_CHAN};
-
-    #[cfg(feature="dma")]
-    use peripheral::dma::{self, DMAChannel, DMA_TX_CHAN4PLUS};
-
-    /// A buffer for transmitting bytes.
-    pub static mut TX_BUFFER: RingBuffer = RingBuffer::new();
-
-    /// A buffer for receiving bytes.
-    pub static mut RX_BUFFER: RingBuffer = RingBuffer::new();
-
-    // Mutex to ensure transmitted data is not jumbled.
-    static WRITE_LOCK: Mutex<()> = Mutex::new(());
-    static READ_LOCK: Mutex<()> = Mutex::new(());
-
-    /// Print a formatted string to the serial port. This macro is intended for
-    /// user code and should not be used to print within the kernel code.
-    #[macro_export]
-    #[cfg(not(test))]
-    macro_rules! print {
-        ($($arg:tt)*) => ({
-            $crate::io::write_fmt(format_args!($($arg)*));
-        });
-    }
-
-    /// Print a formatted string, with a new line appended to it, to the serial port.
-    /// This macro is intended for user code and should not be used to print within
-    /// the kernel code.
-    #[macro_export]
-    #[cfg(not(test))]
-    macro_rules! println {
-        ($fmt:expr) => (print!(concat!($fmt, "\n")));
-        ($fmt:expr, $($arg:tt)*) => (print!(concat!($fmt, "\n"), $($arg)*));
-    }
-
-    struct DMASerial {
-        usart: Usart,
-    }
-
-    impl DMASerial {
-        fn new(usart: Usart) -> Self {
-            DMASerial { usart: usart }
-        }
-    }
-
-    #[cfg(feature="dma")]
-    impl Write for DMASerial {
-        fn write_str(&mut self, string: &str) -> fmt::Result {
-            use peripheral::usart::defs::*;
-
-            let g = CriticalSection::begin();
-            dma::set_dma_usart_tx(DMAChannel::Four,
-                              unsafe {USART2_ADDR.offset((TDR_OFFSET/4) as isize)},
-                              string.as_bytes());
-
-            ::altos_core::syscall::sys_sleep(DMA_TX_CHAN4PLUS);
-            drop(g);
-            Ok(())
-        }
-    }
-
-
-    struct Serial {
-        usart: Usart,
-    }
-
-    impl Serial {
-        fn new(usart: Usart) -> Self {
-            Serial { usart: usart }
-        }
-
-        fn buffer_byte(&mut self, byte: u8) {
-            unsafe {
-                while !TX_BUFFER.insert(byte) {
-                    let _g = CriticalSection::begin();
-                    self.usart.enable_transmit_interrupt();
-                    ::altos_core::syscall::sys_sleep(USART2_TX_CHAN);
-                }
-            }
-        }
-
-        fn read(&mut self, buf: &mut [u8]) -> usize {
-            // UNSAFE: Accessing mutable static
-            while unsafe {
-                let _g = CriticalSection::begin();
-                Volatile::new(&RX_BUFFER).is_empty()
-            } {
-                sleep(USART2_RX_CHAN);
-            }
-            let mut read = 0;
-            while read < buf.len() {
-                let g = CriticalSection::begin();
-                // UNSAFE: Accessing mutable static
-                let byte = unsafe { RX_BUFFER.remove() };
-                drop(g);
-                match byte {
-                    Some(byte) => {
-                        buf[read] = byte;
-                        read += 1;
-                    },
-                    None => break,
-                }
-            }
-            read
-        }
-    }
-
-    impl Write for Serial {
-        fn write_str(&mut self, string: &str) -> fmt::Result {
-            for byte in string.as_bytes() {
-                if *byte == b'\n' {
-                    self.buffer_byte(b'\r');
-                }
-                #[cfg(feature="minicom")]
-                {
-                    if *byte == b'\r' {
-                        self.buffer_byte(b'\n');
-                    }
-                }
-                self.buffer_byte(*byte);
-            }
-            let g = CriticalSection::begin();
-            self.usart.enable_transmit_interrupt();
-            ::altos_core::syscall::sys_sleep(USART2_TX_CHAN);
-            drop(g);
-            Ok(())
-        }
-    }
-
-    struct DebugSerial {
-        usart: Usart,
-    }
-
-    impl DebugSerial {
-        fn new(usart: Usart) -> Self {
-            DebugSerial { usart: usart }
-        }
-
-        fn write_byte(&mut self, byte: u8) {
-            while !self.usart.is_tx_reg_empty() {}
-            self.usart.transmit_byte(byte);
-        }
-    }
-
-    impl Write for DebugSerial {
-        fn write_str(&mut self, string: &str) -> fmt::Result {
-            for byte in string.as_bytes() {
-                // If at end of line, write a carriage return because
-                // minicom doesn't go to beginning of line on its own.
-                if *byte == b'\n' {
-                    self.write_byte(b'\r');
-                }
-                self.write_byte(*byte);
-            }
-            Ok(())
-        }
-    }
-
-    #[doc(hidden)]
-    pub fn write_fmt(args: Arguments) {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut serial = Serial::new(usart2);
-
-        let _g = WRITE_LOCK.lock();
-        serial.write_fmt(args).ok();
-    }
-
-    #[doc(hidden)]
-    pub fn write_str(s: &str) {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut serial = Serial::new(usart2);
-
-        let _g = WRITE_LOCK.lock();
-        serial.write_str(s).ok();
-    }
-
-    #[doc(hidden)]
-    #[cfg(feature="dma")]
-    pub fn dma_write_str(s: &str) {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut dma_serial = DMASerial::new(usart2);
-
-        let _g = WRITE_LOCK.lock();
-        dma_serial.write_str(s).ok();
-    }
-
-    // NOTE: debug assumes interrupts are turned off, so does not need lock.
-    #[no_mangle]
-    #[doc(hidden)]
-    pub fn debug_fmt(args: Arguments) {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut serial = DebugSerial::new(usart2);
-
-        serial.write_fmt(args).ok();
-    }
-
-    // NOTE: debug assumes interrupts are turned off, so does not need lock.
-    #[no_mangle]
-    #[doc(hidden)]
-    pub fn debug_str(s: &str) {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut serial = DebugSerial::new(usart2);
-
-        serial.write_str(s).ok();
-    }
-
-    #[doc(hidden)]
-    pub fn poll_char() -> Option<u8> {
-        let usart2 = Usart::new(UsartX::Usart2);
-        let mut serial = Serial::new(usart2);
-        let mut buf: [u8; 1] = [0];
-        let _g = READ_LOCK.lock();
-        match serial.read(&mut buf) {
-            0 => None,
-            _ => Some(buf[0]),
-        }
-    }
-}
